Move migrations from pfs to pachd main
